pyproject.toml
[tool.poetry]
name = "backend"
version = "1.0.0"
description = ""
authors = ["Dickson Chibuzor"]

[tool.poetry.dependencies]
python = "3.9.5"
# Django
Django = "==2.2.18"
django-extensions = "==3.1.0"
# fields
shortuuid = "==1.0.1"
# postgresql
"psycopg2-binary" = "==2.8.6"
# Rest Framework
djangorestframework = "==3.12.2"
coreapi = "==2.3.3"
pyjwt = "==1.7.1"
django-rest-knox = {git = "https://github.com/Seedstars/django-rest-knox.git", rev = "753fe166060f9a6745209d11d0b305daf19c6a4d"}
django-rest-swagger = "==2.2.0"
drf-spectacular = "==0.17.2"
# Celery
celery = "==5.0.2"
celery-once = "==3.0.1"
pytz = "==2020.5"
redis = "==3.4.1"
django-redis = "==4.11.0"
django-storages = "==1.9.1"
boto3 = "==1.12.45"
feedparser = "==6.0.4"
# Call external REST apis
requests = "==2.25.1"
# Logging
beautifulsoup4 = "==4.9.0"
lxml = "==4.6.2"
cryptography = "==3.3.1"
gunicorn = "==20.0.4"
django-cors-headers = "==3.7.0"

[tool.poetry.dev-dependencies]
responses = "==0.12.1"
factory_boy = "==3.2.0"
prospector = "==1.3.1"
astroid = "==2.4.1"
bandit = "==1.6.2"
pytest = "==6.2.1"
pytest-cov = "==2.10.1"
pytest-django = "==4.1.0"
pytest-pythonpath = "==0.7.3"
pytest-xdist = "==2.2.0"
pytest-socket = "==0.3.5"
ptpython = "==3.0.9"
pylint = "==2.5.3"
pylint-django = "==2.1.0"
pylint-celery = "==0.3"
mock = "==4.0.3"
black = "==19.10b0"
ipython = "==7.19.0"
django-webtest = "==1.9.7"
mypy = "==0.800"
django-stubs = "==1.7.0"
djangorestframework-stubs = "==1.3.0"
moto = "==1.3.17.dev276"
freezegun = "==1.1.0"
flake8 = "==3.8.4"

[build-system]
requires = ["poetry>=0.12"]
